+++ packages/react-ui/storybook/stories/widgetsUI/PieWidgetUI.stories.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PieWidgetUI from '../../../src/widgets/PieWidgetUI';
+import { buildReactPropsAsString } from '../../utils';
 
-// This default export determines where your story goes in the story list
 const options = {
-  title: 'Widgets/PieWidgetUI',
+  title: 'Widgets UI/PieWidgetUI',
   component: PieWidgetUI
 };
 
@@ -17,10 +17,12 @@ const dataDefault = [
 const Template = (args) => <PieWidgetUI {...args} />;
 
 export const Default = Template.bind({});
-Default.args = { data: dataDefault };
+const DefaultProps = { data: dataDefault };
+Default.args = DefaultProps;
+Default.parameters = buildReactPropsAsString(DefaultProps, 'PieWidgetUI');
 
 export const CustomColors = Template.bind({});
-CustomColors.args = {
+const CustomColorsProps = {
   data: [
     { name: 'Dogs', value: 100 },
     { name: 'Cats', value: 120 },
@@ -48,9 +50,11 @@ CustomColors.args = {
     '#7C7C7C'
   ]
 };
+CustomColors.args = CustomColorsProps;
+CustomColors.parameters = buildReactPropsAsString(CustomColorsProps, 'PieWidgetUI');
 
 export const SelectedCategories = Template.bind({});
-SelectedCategories.args = {
+const SelectedCategoriesProps = {
   data: [
     { name: 'Dogs', value: 100 },
     { name: 'Cats', value: 120 },
@@ -61,3 +65,8 @@ SelectedCategories.args = {
   selectedCategories: ['Cats', 'Canaries'],
   onSelectedCategoriesChange: (categories) => console.log(categories)
 };
+SelectedCategories.args = SelectedCategoriesProps;
+SelectedCategories.parameters = buildReactPropsAsString(
+  SelectedCategoriesProps,
+  'PieWidgetUI'
+);